refactor: extract controller from mod and rename it
